moved processing of event to modifier module added nested content to end to end test
